hw17. Corrections based on github conversations.
